api/requirements.txt
Flask
Flask-SQLAlchemy
waitress
requests
SQLAlchemy
python-dotenv
Werkzeug
pymysql
mysql-connector-python